pretty major refactoring, please don't break my legs
So, the idea here is that we don't need the multi pass transmogrifies
anymore. Instead we're back to a variable map, but it's much more
sensible now. The key realization is that the only thing we only ever
print is variable assignments. So the variable map now simply tracks
all of the variable assignments on a given line, and then can print
that out. It also handles iteration, laying the real groundwork for
that stuff.